fix emscripten tests
