+++ lib/emscripten-tests/Cargo.toml
@@ -26,4 +26,3 @@ glob = "0.2.11"
 clif = []
 llvm = ["wasmer-llvm-backend"]
 singlepass = ["wasmer-singlepass-backend"]
-debug = ["wasmer-clif-backend/debug", "wasmer-runtime-core/debug"]
