+++ lib/emscripten-tests/src/tests/is_emscripten_false.wast
